+++ src/webview/html/contentHtmlFromPieceDescriptors.js
